Get Product from Catalog

+++ shipping/Shipping/Domain/Shipping/Price.cs
@@ -2,16 +2,16 @@ using Shipping.Models;
 namespace Shipping.Domain.Shipping;
 public class Price : ValueObject
 {
-    private readonly float _value;
+    private readonly double _value;
 
-    public static Result<Price> Create(float priceValue)
+    public static Result<Price> Create(double priceValue)
     {
         if (priceValue > 0)
             return Result.Ok<Price>(new Price(priceValue));
         return Result.Fail<Price>("Price cannot be a negative value.");
     }
 
-    protected Price(float value)
+    protected Price(double value)
     {
         _value = value;
     }
@@ -20,6 +20,6 @@ public class Price : ValueObject
     {
         yield return _value;
     }
-    public static implicit operator float(Price PriceCode) => PriceCode._value;
-    public static explicit operator Price(float str) => new Price(str);
+    public static implicit operator double(Price PriceCode) => PriceCode._value;
+    public static explicit operator Price(double str) => new Price(str);
 }
